chore: Fix un-related leak test failing on CI due to previous structure
