--- src/Uno.UI.Runtime.Skia.AppleUIKit/UI/Xaml/InputLayer.cs
@@ -1,40 +1,0 @@
-﻿using System;
-using System.Linq;
-using Foundation;
-using UIKit;
-
-namespace Uno.UI.Runtime.Skia.AppleUIKit;
-
-public class InputLayer : UIView
-{
-#if __IOS__
-	public InputLayer()
-	{
-		MultipleTouchEnabled = true;
-	}
-#endif
-
-	public override void TouchesBegan(NSSet touches, UIEvent? evt)
-	{
-		AppleUIKitCorePointerInputSource.Instance.TouchesBegan(this, touches, evt);
-		base.TouchesBegan(touches, evt);
-	}
-
-	public override void TouchesMoved(NSSet touches, UIEvent? evt)
-	{
-		AppleUIKitCorePointerInputSource.Instance.TouchesMoved(this, touches, evt);
-		base.TouchesMoved(touches, evt);
-	}
-
-	public override void TouchesEnded(NSSet touches, UIEvent? evt)
-	{
-		AppleUIKitCorePointerInputSource.Instance.TouchesEnded(this, touches, evt);
-		base.TouchesEnded(touches, evt);
-	}
-
-	public override void TouchesCancelled(NSSet touches, UIEvent? evt)
-	{
-		AppleUIKitCorePointerInputSource.Instance.TouchesCancelled(this, touches, evt);
-		base.TouchesCancelled(touches, evt);
-	}
-}
