+++ src/Uno.UI.Runtime.Skia.AppleUIKit/UI/Xaml/NativeOverlayLayer.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Foundation;
 using UIKit;
+using Uno.UI.Runtime.Skia.AppleUIKit;
 
 namespace Uno.WinUI.Runtime.Skia.AppleUIKit.UI.Xaml;
 
@@ -17,6 +18,8 @@ internal class NativeOverlayLayer : UIView
 		MultipleTouchEnabled = true;
 	}
 #endif
+
+	public event EventHandler? SubviewsChanged;
 
 	public override void SubviewAdded(UIView uiview)
 	{
@@ -36,5 +39,29 @@ internal class NativeOverlayLayer : UIView
 		SubviewsChanged?.Invoke(this, EventArgs.Empty);
 	}
 
-	public event EventHandler? SubviewsChanged;
+	// Note: The pointers are listen here in the NativeOverlayLayer as it's the topmost layer in the view hierarchy.
+	// Note2: This must be in a layer (compared to override RootViewController.TouchesXXX methods) in order to be able to properly get the multitouch events.
+	public override void TouchesBegan(NSSet touches, UIEvent? evt)
+	{
+		AppleUIKitCorePointerInputSource.Instance.TouchesBegan(this, touches, evt);
+		base.TouchesBegan(touches, evt);
+	}
+
+	public override void TouchesMoved(NSSet touches, UIEvent? evt)
+	{
+		AppleUIKitCorePointerInputSource.Instance.TouchesMoved(this, touches, evt);
+		base.TouchesMoved(touches, evt);
+	}
+
+	public override void TouchesEnded(NSSet touches, UIEvent? evt)
+	{
+		AppleUIKitCorePointerInputSource.Instance.TouchesEnded(this, touches, evt);
+		base.TouchesEnded(touches, evt);
+	}
+
+	public override void TouchesCancelled(NSSet touches, UIEvent? evt)
+	{
+		AppleUIKitCorePointerInputSource.Instance.TouchesCancelled(this, touches, evt);
+		base.TouchesCancelled(touches, evt);
+	}
 }
